solved problem with all players could erase the draw, now only the one is drawing can

+++ src/org/academiadecodigo/haltistas/client/Client.java
@@ -122,6 +122,9 @@ public class Client {
     }
 
     public void eraseDraw() {
+        if (!mouseController.isCanDraw()) {
+            return;
+        }
         pencil.deleteAll();
     }
